refactor(ico): split sharp rasterization into raster.ts
