+++ .gitignore
@@ -161,3 +161,4 @@ dist
 AGENTS.md
 /_*.md
 /.idea/
+superpowers/
