+++ package.json
@@ -34,6 +34,7 @@
 		"#types": "./src/types.ts",
 		"#html": "./src/html.ts",
 		"#ico": "./src/ico.ts",
+		"#raster": "./src/raster.ts",
 		"#cli/*": "./src/cli/*.ts",
 		"#cli": "./src/cli.ts",
 		"#instrumentation": "./src/instrumentation.ts",
