+++ src/cli/commands/generate.ts
@@ -2,7 +2,8 @@ import { source } from '#cli/args/source';
 import { blue, green, red } from '#cli/colors';
 import { pathFlag } from '#cli/flags/path';
 import { sizesFlag } from '#cli/flags/sizes';
-import { generateSizedPngs, packIco } from '#ico';
+import { packIco } from '#ico';
+import { generateSizedPngs } from '#raster';
 import { inputBasename, loadInputBytes } from '#loadInput';
 import { command, flag } from '@kjanat/dreamcli';
 import { mkdir, writeFile } from 'node:fs/promises';
